requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
google-api-python-client
google-auth